Add files via upload
The yellow floor on the lower platform was intended to be spikes, but I don't know how to make continuous damage when standing on it or knock-back. The Red Obstacles suffice for environmental damage. Enemies take at least 4 shots to die. You have a health bar to show much health you got left before you kick the bucket.