Read Description:
Working features are as follows: 
The player can reload back to checkpoint with the 'R' key. 
The player will also respawn after death. The player will receive a buff if near proximity of a diamond shaped item. 
The buff gives infinite jumps as long as the player is near.
Floor is lava, burns player if standing too long.
Elevators only elevate when player is on platform.

Visual Representation:
https://www.youtube.com/watch?v=RxMflIV7Ho4