Add files via upload
I managed to partly complete both tasks. I have enemies that have a raycast that is used to spot the player and shoot if player reaches a certain distance. Unfortunately, I did not know how to make shoot where I wanted to causing multiple bullets to spawn and destroy themselves. I use a raycast on the camera that uses the right mouse button that spawns it but only where the camera icon is located in unity.