Add files via upload
I managed to get most of the tasks done within the lab assignment. The biggest problems I ran into I did not know how to fix was first the sprinting of the player to turn back on. I could either get him to jump or sprint but not both. Second I ran into problems with the extra credit, I managed to get it partially when loading from the menu to the game but not back to menu.(back to menu scene works by its self but not when clicking  the quit button from the pause menu in the samplescene.). Third I managed to have some of the option buttons in the sound test menu to print text to the console to show that it can work but had no understanding of how to get it the way you requested.