Marcus Valdivia
Added code in animator and player scripts, as well as functions in state chart for jumping, running idle
Created a pause menu, can be activated when in game when pressing escape. This is created in the Game scene and utilizes my Pause Menu script. 
Used the given main menu canvas. Created code for SFX slider and BGM slider. Each SFX and BGM have their own sound mixer, both with the attached appropriate audio sources. This script is the SoundSlider.cs